app_streamlit/requirements.txt
streamlit
pandas
plotly
openpyxl
python-dotenv
s3fs
SQLAlchemy
psycopg2-binary